requirements.txt
streamlit
faiss-cpu
sentence-transformers
openai
pypdf
python-docx
pandas
numpy
tqdm
